+++ src/neo4j/_async/work/workspace.py
@@ -22,7 +22,7 @@ import typing as t
 from ..._async_compat.util import AsyncUtil
 from ..._auth_management import to_auth_dict
 from ..._conf import WorkspaceConfig
-from ..._meta import (
+from ..._warnings import (
     deprecation_warn,
     unclosed_resource_warn,
 )